requirements.txt
aiogram==2.23.1
aiohttp==3.8.3
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
debugpy==1.6.5
Django==4.1.4
django-admin-argon-dashboard==1.0.16
django-autocomplete-light==3.9.4
django-filter==22.1
django-model-utils==4.3.1
django-money==3.0.0
django-phonenumber-field==7.0.1
django-phonenumbers==1.0.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
importlib-metadata==5.2.0
magic-filter==1.0.9
Markdown==3.4.1
multidict==6.0.4
phonenumbers==8.13.3
py-moneyed==2.0
pycodestyle==2.10.0
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.8.2
zipp==3.11.0
